int32 -> int64 to get correct casting

+++ manet/data/mammo_data.py
@@ -96,7 +96,7 @@ class MammoDataset(Dataset):
         label_fn = self.data_root / data_dict['label_fn']
 
         mammogram = read_mammogram(image_fn)
-        mask = read_image(label_fn, force_2d=True, no_metadata=True, dtype=np.uint8)
+        mask = read_image(label_fn, force_2d=True, no_metadata=True, dtype=np.int64)  # int64 gets cast to LongTensor
 
         sample = {
             'mammogram': mammogram,
